remove unneccessary useEffect dependencies
